+++ src/plugins/llm/README.md
@@ -282,13 +282,14 @@ LLM__TTS_MODEL=default
 
 DeepSeek 字段与鉴权方式以[官方余额接口文档](https://api-docs.deepseek.com/zh-cn/api/get-user-balance)为准。
 
-| Provider   | 追加路径        | 说明                                                           |
+| Provider   | 接口根路径      | 说明                                                           |
 | ---------- | --------------- | -------------------------------------------------------------- |
 | `deepseek` | `/user/balance` | 请求 DeepSeek 官方余额接口；默认复用模型 `api_key`             |
 | `aperture` | `/api/quotas`   | 请求 Tailscale Aperture 额度接口；可用 `bucket` 筛选单个额度桶 |
 
-未设置 `quota.api_url` 时，插件优先使用 `LLM__BASE_URL` 并追加上表路径；全局地址也为空时，
-再使用模型解析后的服务地址。`quota.api_url` 可填写完整地址以覆盖这一行为。两种 provider 还可单独设置
+未设置 `quota.api_url` 时，插件优先使用 `LLM__BASE_URL` 的域名根目录解析上表路径；例如
+`https://ai.example.com/v1` 会解析为 `https://ai.example.com/api/quotas`。全局地址为空时，再使用模型解析后的
+服务地址。额度接口位于其他域名或非标准路径时，可用完整的 `quota.api_url` 覆盖。两种 Provider 还可单独设置
 `api_key`、`proxy` 与 `timeout`。
 
 ## 内置工具
